Fix format specifier warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36474 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/gfxoutputdrv/godotdrv.c
@@ -63,7 +63,10 @@ static int godotdrv_write_file_header(screenshot_t *screenshot)
         }
         return 0;
     }
-    if (fprintf(fd, "GOD1%c%c%c%c", 0, 0, screenshot->width / 8, screenshot->height / 8) < 0) {
+    if (fprintf(fd, "GOD1%c%c%c%c",
+                0, 0,
+                (int)(screenshot->width / 8),
+                (int)(screenshot->height / 8)) < 0) {
         return -1;
     }
     return 0;
